backend/pyproject.toml
[project]
name = "huerai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "pydantic-settings>=2.13.1",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
    "psycopg2-binary>=2.9",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "anthropic>=0.40",
]
